requirements.txt
pyjwt==2.1.0
flask==2.0.1
gunicorn==20.1.0
pytest==6.2.4
